requirements.txt
pandas
numpy
matplotlib
jupyter
azure-storage-blob
python-dotenv
